requirements.txt
albumentations==2.0.8
# datasets==3.6.0
# evaluate==0.4.5
# huggingface_hub==0.34.4
matplotlib==3.10.5
numpy==2.2.6
opencv_python
opencv_python_headless
# optuna==4.4.0
pandas==2.3.1
# peft==0.17.0
Pillow==11.3.0
psutil==7.0.0
scikit_learn==1.7.1
scipy==1.15.3
simpleitk==2.5.0
timm==1.0.19
torch==2.7.1
torchvision==0.22.1
trl==0.21.0
wandb==0.21.3
psrcal== 1.0.0
lightning==2.4.0
omegaconf==2.3.0